requirement.txt
numpy==1.14.2
opencv-contrib-python==3.4.0.12
Pillow==5.1.0
progressbar33==2.4
PyYAML==3.12
six==1.11.0
torch==0.3.1
torchvision==0.2.0
